+++ app/src/main/java/fathorazi/id/ac/unuja/bus/SplaceScreenWSActivity.java
@@ -6,10 +6,10 @@ import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
 
 public class SplaceScreenWSActivity extends AppCompatActivity {
-    /* access modifiers changed from: protected */
+
     public void onCreate(Bundle bundle) {
         super.onCreate(bundle);
-        setContentView((int) R.layout.activity_splace_screen_ws);
+        setContentView(R.layout.activity_splace_screen_ws);
         new Thread() {
             public void run() {
                 try {
@@ -22,7 +22,6 @@ public class SplaceScreenWSActivity extends AppCompatActivity {
         }.start();
     }
 
-    /* access modifiers changed from: protected */
     public void onDestroy() {
         super.onDestroy();
     }
